Refactor NCA shaders to use f32 for calculations, improve living state handling, and enhance optimizer with f32 master parameters
- Updated nca_backward_f16.wgsl and nca_forward_f16.wgsl to replace f16 with f32 for various calculations, improving precision and performance.
- Introduced living_maxima function to determine living state based on thresholds in nca_forward_f16.wgsl.
- Modified optimizer_f16.wgsl to maintain master parameters and moments in f32, ensuring stability in Adam optimization.
- Added pool_recovery_f16.wgsl for handling state recovery from a pool, including checks for living states.
- Enhanced CSS styles for better UI consistency and responsiveness.
- Updated notebook to reset execution count for reproducibility.

+++ demo/train_growing_voxels/shaders/dense_backward_f16.wgsl
@@ -25,17 +25,17 @@ fn activation_delta(@builtin(global_invocation_id) gid: vec3<u32>) {
   if (index >= total) { return; }
   let row = index / params.output_dim;
   let output_channel = index - row * params.output_dim;
-  var preactivation = biases[output_channel];
+  var preactivation = f32(biases[output_channel]);
   for (var input_channel = 0u; input_channel < params.input_dim; input_channel++) {
-    preactivation += input_values[row * params.input_dim + input_channel]
-      * weights[input_channel * params.output_dim + output_channel];
+    preactivation += f32(input_values[row * params.input_dim + input_channel])
+      * f32(weights[input_channel * params.output_dim + output_channel]);
   }
   let derivative = select(
-    f16(1.0),
-    f16(params.omega) * cos(f16(params.omega) * preactivation),
+    1.0,
+    params.omega * cos(params.omega * preactivation),
     params.omega > 0.0,
   );
-  output_values[index] = upstream_values[index] * derivative;
+  output_values[index] = f16(f32(upstream_values[index]) * derivative);
 }
 
 @compute @workgroup_size(64)
@@ -45,12 +45,12 @@ fn dense_dx(@builtin(global_invocation_id) gid: vec3<u32>) {
   if (index >= total) { return; }
   let row = index / params.input_dim;
   let input_channel = index - row * params.input_dim;
-  var sum = f16(0.0);
+  var sum = f32(0.0);
   for (var output_channel = 0u; output_channel < params.output_dim; output_channel++) {
-    sum += input_values[row * params.output_dim + output_channel]
-      * weights[input_channel * params.output_dim + output_channel];
+    sum += f32(input_values[row * params.output_dim + output_channel])
+      * f32(weights[input_channel * params.output_dim + output_channel]);
   }
-  output_values[index] = sum;
+  output_values[index] = f16(sum);
 }
 
 @compute @workgroup_size(64)
@@ -60,21 +60,21 @@ fn dense_dw(@builtin(global_invocation_id) gid: vec3<u32>) {
   if (index >= total) { return; }
   let input_channel = index / params.output_dim;
   let output_channel = index - input_channel * params.output_dim;
-  var sum = f16(0.0);
+  var sum = f32(0.0);
   for (var row = 0u; row < params.rows; row++) {
-    sum += input_values[row * params.input_dim + input_channel]
-      * upstream_values[row * params.output_dim + output_channel];
+    sum += f32(input_values[row * params.input_dim + input_channel])
+      * f32(upstream_values[row * params.output_dim + output_channel]);
   }
-  output_values[index] = sum;
+  output_values[index] = f16(sum);
 }
 
 @compute @workgroup_size(64)
 fn dense_db(@builtin(global_invocation_id) gid: vec3<u32>) {
   let output_channel = gid.x;
   if (output_channel >= params.output_dim) { return; }
-  var sum = f16(0.0);
+  var sum = f32(0.0);
   for (var row = 0u; row < params.rows; row++) {
-    sum += upstream_values[row * params.output_dim + output_channel];
+    sum += f32(upstream_values[row * params.output_dim + output_channel]);
   }
-  output_values[output_channel] = sum;
+  output_values[output_channel] = f16(sum);
 }
